23-line-chatbot-for-chat-with-data/requirements.txt
functions-framework==3.8.2
line-bot-sdk==3.13.0
google-cloud-discoveryengine==0.12.3
google-cloud-storage==2.18.2
vertexai==1.71.1
google-cloud-dialogflow-cx==1.36.0

